+++ finalProject/src/main/java/com/study/finalProject/repository/StudyRepository.java
@@ -1,6 +1,8 @@
 package com.study.finalProject.repository;
 
 import java.util.List;
+import java.util.Optional;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -28,4 +30,6 @@ public interface StudyRepository extends JpaRepository<Study, Long>, JpaSpecific
 
     // 특정 pid와 연관된 모든 Study 반환
     List<Study> findByPatient_Pid(String pid);
+    
+    Optional<Study> findById(Long studyKey);
 }
