+++ finalProject/src/main/resources/templates/patientDetail.html
@@ -10,12 +10,11 @@
         <h1>상세정보</h1>
         <button class="edit-button" id="editBtn">수정버튼</button>
         <div class="patient-info">
-            <p>환자 이름: <span th:text="${patient.pName}">홍길동</span></p>
+            <!--<p>환자 이름: <span th:text="${patient.pName}">홍길동</span></p>-->
             <p>환자 ID: <span th:text="${patient.pid}">USER01</span></p>
             <p>생일: <span th:text="${patient.pBirthDate}">2000-12-31</span></p>
             <p>성별: <span th:text="${patient.pSex}">남</span></p>
             <p>코멘트: <span th:text="${patient.comments}">환자 상태 양호</span></p>
-            <button type="button" onclick="location.href='/reportPage?pid=' + /*[[${patient.pid}]]*/ 'USER01';">리포트</button>
         </div>
 
         <!-- 코멘트 입력 폼 -->
